Misc: Preprocessor goes brrrr
These were all redundant. The preprocessor is perfectly capable of remapping
between Forge/Fabric unless forge touched something (which is not the case
here).

One thing to note is that for UTextComponent, two different implementations were
used because it seemed like Fabric's text class is immutable whereas Forge's is
not. In reality they both are identical and immutable and Forge just has
horrible method names which look like setters but return a new object.
So, the Fabric implementation was preserved (just remapped to Forge names) cause
the Forge one was fundamentally broken.