Build: Temporarily disable Fabric sub-project
Eclipse did a breaking change (bump from Java 8 to 11) without a major version
bump and is using fuzzy versions, so we automatically get the incompatible one
and CI breaks.
As a workaround, this commit disables the fabric sub-project until we switch to
Java 16 ourselves (that is, until we merge the `next` branch).