Misc: Do not use typealiases where not necessary
I believe this started when the dumb preprocessor was still in use but it does
not make much sense now that it can remap everything by itself.

Additionally, using typealias actually causes issues for third-parties using
different mappings in development than we do because Kotlin's metadata (which is
the only place typealiases exist in) is not remapped and the Kotlin compiler
depends on it, leading to compilation failures because it cannot find the
classes.

The only aliases which we do still use are the Action ones because of the
generic. Ideally we re-write UTextComponent to take those into account. The
current implementation is very questionable.