Build: Remove unnecessary mappings and preprocessor statements
The preprocessor can infer all of those mappings perfectly fine by itself and
even the manual preprocessor comment are unnecessary (presumable it got broken
because of the manual mappings).
Tested by looking at the diff between the 1.18-forge source before and after
these changes and the only change in the diff is the removal of the preprocessor
comments, i.e. no semantic difference.