UScreen: Rename `name` to `unlocalizedName`
So it is clear to users that they should pass a translation key into the
constructor and that they need to pass the value through I18n before displaying
it to the user.
This is technically a breaking change but the previous version was merged
without approval and had not yet been deployed, so it should be fine to change.