UGraphics: Add CommonVertexFormats
The VertexFormats class which is currently being used with all the begin methods
is from Minecraft, meaning it won't be available to version-independent code.
There's nothing wrong though with version-independent code trying to render
primitives, so to facilitate that use case, we provide a wrapper enum for all
vanilla vertex formats which exist on all versions.

Upon re-generating the api file, two version-dependent entries were moved. These
are still present, the merging algorithm is just not perfect and may change the
order of version-dependent entries when surrounding entries change.