UMatrixStack: Add version-independent model matrix accessor
UniversalCraft generally makes for a pretty good abstraction of a significant
chunk of Minecraft, such that one can write code which is only compiled once but
works with all versions of Minecraft.
Such code however cannot access the model matrix of a UMatrixStack because its
type is `Matrix4f` which is a Minecraft type that changes across version.
This commit provides an accessor to retrieve the content of the matrix in a
version independent way (plain old FloatArray).