UGraphics: Fix incorrect default shader being used
We store default shaders for each built-in vertex format in a `HashMap`. Turns
out though that `VertexFormat`'s `equals` implementation is broken and returns
true for two `VertexFormat`s that use the same elements but in a different
order (such as PosTexColorLight vs PosTexLightColor). As a result if you tried
to draw the former, it'd actually use the shader for the latter and misinterpret
your light texture coordinates as the vertex color and vice versa, usually
producing invalid or simply no visible output.