Fix gui scale handling on 1.20.5
On 1.20.5, the change callback of the guiScale option calls
MinecraftClient#onResolutionChanged, which in turn causes the current
screen to be reinitialized. However, UScreen calls updateGuiScale when
the screen is initialized, causing the gui scale to be set back to
the overridden gui scale. This doesn't recurse forever, since Minecraft
only calls option change callbacks when the value changes.

We fix this by only updating the gui scale when guiScaleToRestore is -1,
and restore this field to -1 when the screen is closed (but after
restoring the gui scale). This also fixes gui scale restoration with
screen instances that are reused, previously the gui scale that was
recorded when the gui was opened for the first time would always be
restored.