Build: Port to 1.20.6 Fabric
* Build: Port to 1.20.6 Fabric

Sets kotlin language and api version to maintain compatibility with
consumers still using older Kotlin compilers / stdlib.

* UScreen: Fix gui scale handling

As of 1.20.5, the change callback of the guiScale option calls
MinecraftClient#onResolutionChanged, which in turn causes the current
screen to be reinitialized. However, UScreen calls updateGuiScale when
the screen is initialized, causing the gui scale to be set back to
the overridden gui scale. This doesn't recurse forever, since Minecraft
only calls option change callbacks when the value changes.

We fix this by ignoring this call when we are the ones that trigger it.

We also restore the guiScaleToRestore field to -1 when the screen is closed.
This fixes gui scale restoration with screen instances that are reused,
previously the gui scale that was recorded when the gui was opened for
the first time would always be restored.

* Build: remove old code from gradlew

This was from when we used TeamCity, the IS_CI variable not set on
GitHub Actions. Java versions are now managed by the GitHub Actions
workflow.

GitHub: #68