Standalone: Fix slf4j logging and add default log4j config
We use version 2 of slf4j, so we need to include the log4j adapter which
was built for version 2, not the one for version 1.

With that, logging works, but it's using log4j's ugly default format and
only emits WARN or above.
To improve the out-of-the-box behavior for Standalone UC, we add a
default log4j config file; applications which require further
customization can override it via one of the many way which log4j
provides.