UScreen: Fix background drawing on 1.21.2+
As of 1.21.2 MC's DrawContext will now by default buffer all drawn quads
and only flush when strictly necessary.
This however requires that everyone is in on the batching and has all
their calls batched as well, otherwise buffered stuff will get rendered
much later than it was supposed to. E.g. the vanilla background (much
more obvious with the Programmer Art pack which brings back the dirt
background) will be rendered on top of most of the Elementa gui.

Ideally we'd get in on the batching too, but that's quite the task, so
for the time being this commit fixes the issue by explictly flushing
after every time we pass the DrawContext to a vanilla method (i.e. any
time something could have been buffered).