UTextComponent: Deprecate
Porting it to 1.21.5 while maintaining its API is difficult (actions are
no longer separate from their values and those are no longer stringly
typed either).

As such, this commit deprecates the class as well as the closely
adjacent (and also badly designed) `UMessage` class.

For methods which are still generally useful, it introduces overloads
which take the vanilla component.
`UChat.chat` was already taking `Any`, so `String` and vanilla
components can simply be passed to it directly.
It also moves out the code which turns the component into a plain
(formatted or unformatted) string into separate utilities because that
is still very much useful (and about the only justified use-case of
UTextComponent in Essential).

The only significant loss is the ability of `UMessage` to edit messages
which have been posted in chat. If anyone has a use-case for these, we
may re-introduce this functionality as a new, better-designed API in the
future.