URenderPipeline: Fix incorrect blend state on 1.17 - 1.20.6
The blend state passed to `UShader.fromLegacyShader`, which we ignore
explicitly via `skipBlendState` on versions prior to 1.17, can still get
applied by MC's shader class on 1.17 - 1.20.6 and we can't easily
disable/overwrite that because it happens far down the draw method call.
As such, we must pass the correct blend state to it, otherwise it'll
sometimes disable blending when we want it enabled.

This does not affect 1.21+ because MC's shader class no longer cares
about the blend state on those versions.

Additionally, this issue is masked by the fact that the caching in MC's
GlBlendState class is horribly broken in multiple ways, so often it
wouldn't actually apply at all, thereby making this issue inconsistent
even on 1.17 - 1.20.6.

Linear: EM-3164