Build: Set `FMLModType` depending on which ModLauncher version is used
This attribute was previously set to `LIBRARY` before 0696479 where it
was removed for the following reason:
> however since we require shading and relocation on ML9 anyway,
  we may as well just remove it outright.

This is fine for production, but relocation does not apply in
development environments. The type should be declared as `GAMELIBRARY`
on >=ML9, and `LIBRARY` on <=ML8.

Anything which uses UniversalCraft should also be on the `GAME` layer,
and cannot be on any other layer (e.g. PLUGIN) as it will not be
able to access UniversalCraft classes. If we want other libraries to
be usable in development environments (Vigilance, Elementa), we'll have
to ship separate artifacts of those for >=ML9 and <=ML8 with the
`FMLModType` attribute set accordingly.