AdvancedDrawContext: Fix memory leak on disconnect
Minecraft clears its task queue on disconnect from a (potentially
integrated) server. If this happens after the task is queued in
TemporaryTextureAllocator but before it is executed, it can get stuck in
a state where the `taskQueued` is true but the task will never actually
run, and consequently we'll never clean up old textures and just keep
allocating new ones, quickly running out of memory.

There doesn't appear to be any other queue we could use instead, so this
commit changes the approach to require the user of `AdvancedDrawContext`
to manually manage its lifetime.

Of note, it doesn't appear to be possible to actually hit this condition
in UC unless one has a custom screen which disconnects from the server,
since the vanilla disconnect button isn't on a UC screen and
server-caused disconnect is queued on the same queue and therefore
happens to always run in such a way that this is not triggered.
The copy of this class in Essential which powers its overlays can
reliably trigger the issue though, since it displays on the ingame menu
where the Disconnect/Save&Quit button is located.

Linear: EM-3293
GitHub: #107