UScreen: Fix broken projection matrix after background rendering
The panorama background renderer sets (and doesn't restore) the global
projection matrix to use perspective projection, so we need to save it
before calling the `renderBackground` method and restore it afterwards.
This isn't a problem for vanilla because it doesn't use the projection
matrix during gui screen drawing (it sets up a new one in GuiRenderer
once it actually commits the draw calls), but our `AdvancedDrawContext`
does.

The is masked with Essential installed because it uses
`AdvancedDrawContext` for its overlays which re-sets a proper projection
matrix (at least right now because it doesn't yet optimize to check if
there is actually anything to display, nor does it capture+restore the
projection matrix, which it probably should).

GitHub: #108