Build: Port to 1.21.7 Fabric
There are no changes required for this, however 1.21.7 adds some new
methods so we can't just drop 1.21.6 and use the 1.21.7 build without
risking us accidentially using those in the future, resulting in
NoSuchMethodError on 1.21.6.
And we need a 1.21.7-fabric build to be able to remap to
1.21.7-(neo)forge.

Also bumps 1.21.6 to compile against the release instead of rc1, and
adds it to the README which I forgot to do in its initial commit.