UGraphics: Fix drawString with custom shadow color and bold font
To render with a custom shadow color, which is not something Minecraft's
builtin shadow rendering supports, we render the same text twice with
the first time offset and colored to appear as the shadow.
To render bold font, Minecraft draws each glyph twice with the second
time offset by 1 horizontally to give the bold effect.

Prior to Minecraft 1.21.6, when no builtin shadow is involved (as is the
case with our custom shadow rendering), both of these glyph rectangles
had the exact same z value, and therefore all rectangles drawn for our
shadow pass and our text pass all had the exact same z values. The
default depth test function is LESS_EQUAL, so all rectangles were drawn
based purely on the draw call order.
As of Minecraft 1.21.6 however, the vanilla text renderer now places the
second rectangle of a bold glpyh every so slightly (0.001) closer to the
camera (presumably to prevent some weird z-fighting when viewed at just
the wrong angles). This means the second rectangle of our shadow pass
will now appear in front of the first rectangle of our regular pass,
giving incorrect results.

This commit fixes the issue by doing what vanilla does (and has done
even prior to 1.21.6, we just didn't need it there) when you use its
builtin shadow rendering: Offset the normal pass to be in front of the
shadow pass. And notably more (0.003) than the second rectangle of a
bold glyph, such that the normal pass is now always completely in front
of the shadow pass.

GitHub: #110
Linear: EM-3301