URenderPipeline: Fix rendering being affected by global glShadeModel state
The `glShadeModel` state is initially `GL_SMOOTH`, but other code may
change it to `GL_FLAT` and if it doesn't change it back afterwards,
users of `URenderPipeline` may not get the expected render result (e.g.
trying to draw a gradient will result in a single flat color instead).

The chance of some user of `URenderPipeline` intentionally relying on
this global state leaking in feels sufficiently low, that I don't think
any backwards compatibility needs to be maintained for it.
I don't think it is even necessary to make this behavior configurable,
since we don't allow a single vertex to be shared by multiple
primitives, so anything which could be achieved by using GL_FLAT can
also by achieved by simply using the same data for all vertices of each
primitive. And modern OpenGL doesn't support it anyway.

GitHub: #116