URenderPipeline: Fix blend state with default/mc shaders on 1.17-1.20.6, third try
There another bug in GlBlendState which cause the GlBlendState which we
apply by enabling the shader to not actually be stored as the
`activeBlendState`.
See code comments for details.

Since we now necessarily have access to the GlBlendState of the shader,
this commit also simplifies our dummy call to merely enable that
particular GlBlendState instead of binding and unbinding the entire
shader.

Linear: EM-3512
GitHub: #121